Moved all serializer related classes into its own folder, to separate the concept from the actual "model" classes.
